Set up component demo pages with one for general components and another for components that require kiva app access.
